Replace Parser with LALRPOP and generate errors with codespan
The older Logos lexer and hand written parser have been scrapped for a parser
generated with LALRPOP. This allows me to make changes quickly without having
to rewrite a million things.

The old parse error reporting system has been replaced with the codespan
crate. It looks a lot better and is more ergonomic to use throughout the AST
and IR.

Signed-off-by: Jadon Fowler <j@jadon.io>