IR: Change alloca to store a reference to the instruction it's allocating for
Instead of holding the type it needs to allocate, alloca holds an InstructionId
for the instruction that holds the expression that is being allocated into the
variable. This means we can check the type that needs to be allocated *after*
the IR conversion phase. Before, this was erroring with declarations that didn't
exist in the Compiler while a declaration with alloca was being generated.

Signed-off-by: Jadon Fowler <j@jadon.io>