LLVM Backend: produce instructions from IR
Blocks have been reworked a bit. They now have a "starting
offset" that acts at the amount of instructions before the block.
This means that blocks in a function now have instructions that
maintain order through the entire function and not just for the
block itself, removing naming conflicts between instructions in
separate blocks.

Signed-off-by: Jadon Fowler <j@jadon.io>